Changed Functors/Monads/Writer folder topology
