+++ src/Functors/Monads/Writer/index.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * writer monad functions
+ *
+ * @package bingo-functional
+ * @author Lochemem Bruno Michael
+ * @license Apache-2.0
+ */
+
+require_once __DIR__ . '/ExecWriter.php';
+require_once __DIR__ . '/MapWriter.php';
+require_once __DIR__ . '/RunWriter.php';
+require_once __DIR__ . '/Tell.php';
+require_once __DIR__ . '/Writer.php';
